refactor: update custom information dir name
