+++ models/organization-extension/src/index.ts
@@ -1,3 +1,5 @@
+export * as ContactInformation from './contact-information';
+
 export { default as random } from './builder';
 export { default as presets } from './presets';
 export * from './types';
